AlternanceCode/requirements.txt
PyYAML>=5.4.0
easydict>=1.9
matplotlib>=3.4.0
medpy>=0.4.0
nibabel>=3.2.0
numpy>=1.21.0
pandas>=1.3.0
sklearn>=0.24.0
torch>=1.12.0
torchvision>=0.13.0
tqdm>=4.60.0
